Finished assignment 2 - My First Distributed System Part 2

+++ distributed_system/src/server/MathLogic.java
@@ -2,12 +2,12 @@ package server;
 
 public class MathLogic {
 
-	public static int add(int a, int b)
+	public static double add(double a, double b)
 	{
 		return a+b;
 	}
 	
-	public static int subtract(int a, int b)
+	public static double subtract(double a, double b)
 	{
 		return a-b;
 	}
